Nivell_1:_JUnit/Exercici_2/CalculoDniTest.java
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

public class CalculoDniTest {

    @Test
    public void testCalcularLetraDni() {
        // Update expected letters based on the correct calculation
        Assertions.assertEquals("12345678T", CalculoDni.calcularLetraDni("12345678")); // Update based on correct calculation
        Assertions.assertEquals("23456789R", CalculoDni.calcularLetraDni("23456789"));
        Assertions.assertEquals("34567890W", CalculoDni.calcularLetraDni("34567890"));
        Assertions.assertEquals("45678901C", CalculoDni.calcularLetraDni("45678901")); // Update as needed
        Assertions.assertEquals("98765432Y", CalculoDni.calcularLetraDni("98765432")); // Ensure this is correct
    }

    @Test
    public void testInvalidDni() {
        Assertions.assertThrows(IllegalArgumentException.class, () -> {
            CalculoDni.calcularLetraDni("InvalidNumber");
        });
        Assertions.assertThrows(IllegalArgumentException.class, () -> {
            CalculoDni.calcularLetraDni("1234567"); // Too short
        });
        Assertions.assertThrows(IllegalArgumentException.class, () -> {
            CalculoDni.calcularLetraDni("123456789"); // Too long
        });
        Assertions.assertThrows(IllegalArgumentException.class, () -> {
            CalculoDni.calcularLetraDni("1234567A"); // Contains non-numeric character
        });
        Assertions.assertThrows(IllegalArgumentException.class, () -> {
            CalculoDni.calcularLetraDni(null); // Null input
        });
    }
}
